Redesigned whole items fetching system and other improvements.
- Redesigned the whole system. Now feeds are analyzed by the databased
and the resources are classified based on base keywords and topic
specific keywords inputted into the database.
- The system support base keywords, topic specific keywords and topic
specific exclusion keywords.
- Implemented a query syntax to input the keywords.
- The system uses FiveFilters.org to get main content of pages to match
keywords against. If the site is not supported correctely by
fivefilters or fivefilters fail, the whole page is taken for matching.
- Implemented to Excluded Resources object to track excluded resources
and reason of esclusion.
- SharedCount.appspot.com is used to retrieve social sharing levels of
resources.
- Fixed glitch that was making new resources showing some empty fields
as non empty on creation.
- Implemented fail safe algorithm to clean resources not saved
correctly due to server hangs.
- Feeds analysis and sharing levels assignements are triggered by
cronjobs